Add custom authorization for restaurant operations

+++ Restaurant.Application/Dishes/Commands/DeleteDish/DeleteDishesFromRestaurantCommandHandler.cs
@@ -1,20 +1,29 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDish;
 
 public class DeleteDishesFromRestaurantCommandHandler(ILogger<DeleteDishesFromRestaurantCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository
+    IDishesRepository dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService
     ) : IRequestHandler<DeleteDishesFromRestaurantCommand>
 {
     public async Task Handle(DeleteDishesFromRestaurantCommand request, CancellationToken cancellationToken)
     {
         logger.LogInformation("Deleting dish with RestaurantId: {RestaurantId}", request.RestaurantId);
         var restaurant = await restaurantsRepository.GetOneByIdAsync(request.RestaurantId) ?? throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+        {
+            throw new ForbidException();
+        }
+
         await dishesRepository.DeleteAsync(restaurant.Dishes);
     }
 }
